+++ src/ssh/azext_ssh/custom.py
@@ -8,11 +8,7 @@ import os
 import hashlib
 import json
 import tempfile
-import urllib.request
-import base64
-import stat
 import time
-from glob import glob
 
 from knack import log
 from azure.cli.core import azclierror
@@ -22,15 +18,14 @@ from azure.core.exceptions import ResourceNotFoundError, HttpResponseError
 from . import ip_utils
 from . import rsa_parser
 from . import ssh_utils
-from . import constants as consts
-from . import file_utils
+from . import arc_utils
 
 logger = log.get_logger(__name__)
 
 
 def ssh_vm(cmd, resource_group_name=None, vm_name=None, ssh_ip=None, public_key_file=None,
            private_key_file=None, use_private_ip=False, local_user=None, cert_file=None, port=None,
-           ssh_client_path=None, delete_credentials=False, resource_type=None, arc_proxy_folder=None, ssh_args=None):
+           ssh_client_path=None, delete_credentials=False, resource_type=None, ssh_proxy_folder=None, ssh_args=None):
 
     if delete_credentials and os.environ.get("AZUREPS_HOST_ENVIRONMENT") != "cloud-shell/1.0":
         raise azclierror.ArgumentUsageError("Can't use --delete-private-key outside an Azure Cloud Shell session.")
@@ -40,13 +35,13 @@ def ssh_vm(cmd, resource_group_name=None, vm_name=None, ssh_ip=None, public_key_
     do_ssh_op = _decide_op_call(cmd, resource_group_name, vm_name, ssh_ip, resource_type, None, None,
                                 ssh_client_path, ssh_args, delete_credentials, credentials_folder, local_user)
     do_ssh_op(cmd, vm_name, resource_group_name, ssh_ip, public_key_file, private_key_file, local_user,
-              cert_file, port, use_private_ip, credentials_folder, arc_proxy_folder)
+              cert_file, port, use_private_ip, credentials_folder, ssh_proxy_folder)
 
 
 def ssh_config(cmd, config_path, resource_group_name=None, vm_name=None, ssh_ip=None,
                public_key_file=None, private_key_file=None, overwrite=False, use_private_ip=False,
                local_user=None, cert_file=None, port=None, resource_type=None, credentials_folder=None,
-               arc_proxy_folder=None):
+               ssh_proxy_folder=None):
 
     if (public_key_file or private_key_file) and credentials_folder:
         raise azclierror.ArgumentUsageError("--keys-destination-folder can't be used in conjunction with "
@@ -70,7 +65,7 @@ def ssh_config(cmd, config_path, resource_group_name=None, vm_name=None, ssh_ip=
     do_ssh_op = _decide_op_call(cmd, resource_group_name, vm_name, ssh_ip, resource_type, config_path, overwrite,
                                 None, None, False, credentials_folder, local_user)
     do_ssh_op(cmd, vm_name, resource_group_name, ssh_ip, public_key_file, private_key_file, local_user,
-              cert_file, port, use_private_ip, credentials_folder, arc_proxy_folder)
+              cert_file, port, use_private_ip, credentials_folder, ssh_proxy_folder)
 
 
 def ssh_cert(cmd, cert_path=None, public_key_file=None):
@@ -86,12 +81,17 @@ def ssh_cert(cmd, cert_path=None, public_key_file=None):
                        "is no longer being used.", keys_folder)
     public_key_file, _, _ = _check_or_create_public_private_files(public_key_file, None, keys_folder)
     cert_file, _ = _get_and_write_certificate(cmd, public_key_file, cert_path)
-    print(cert_file + "\n")
+    try:
+        certificate_lifetime = ssh_utils._get_certificate_lifetime(cert_file)
+        print(f"Generated SSH certificate {cert_file} is valid for {certificate_lifetime}.")
+    except Exception as e:
+        logger.warning("Couldn't determine certificate validity. Error: %s", str(e))
+        print(cert_file + "\n")
 
 
 def ssh_arc(cmd, resource_group_name=None, vm_name=None, public_key_file=None, private_key_file=None,
             local_user=None, cert_file=None, port=None, ssh_client_path=None, delete_credentials=False,
-            arc_proxy_folder=None, ssh_args=None):
+            ssh_proxy_folder=None, ssh_args=None):
 
     if delete_credentials and os.environ.get("AZUREPS_HOST_ENVIRONMENT") != "cloud-shell/1.0":
         raise azclierror.ArgumentUsageError("Can't use --delete-private-key outside an Azure Cloud Shell session.")
@@ -118,19 +118,19 @@ def ssh_arc(cmd, resource_group_name=None, vm_name=None, public_key_file=None, p
     op_call = functools.partial(ssh_utils.start_ssh_connection, ssh_client_path=ssh_client_path, ssh_args=ssh_args,
                                 delete_credentials=delete_credentials)
     _do_ssh_op(cmd, vm_name, resource_group_name, None, public_key_file, private_key_file, local_user, cert_file, port,
-               False, credentials_folder, arc_proxy_folder, op_call, True)
+               False, credentials_folder, ssh_proxy_folder, op_call, True)
 
 
 def _do_ssh_op(cmd, vm_name, resource_group_name, ssh_ip, public_key_file, private_key_file, username,
-               cert_file, port, use_private_ip, credentials_folder, arc_proxy_folder, op_call, is_arc):
+               cert_file, port, use_private_ip, credentials_folder, ssh_proxy_folder, op_call, is_arc):
 
-    if not is_arc and arc_proxy_folder:
-        logger.warning("Target machine is not an Arc Server, --arc-proxy-folder value will be ignored.")
+    if not is_arc and ssh_proxy_folder:
+        logger.warning("Target machine is not an Arc Server, --ssh-proxy-folder value will be ignored.")
 
     # If user provides local user, no credentials should be deleted.
     delete_keys = False
     delete_cert = False
-    certificate_validity_in_seconds = None
+    cert_lifetime = None
     if not username:
         delete_cert = True
         public_key_file, private_key_file, delete_keys = _check_or_create_public_private_files(public_key_file,
@@ -139,16 +139,16 @@ def _do_ssh_op(cmd, vm_name, resource_group_name, ssh_ip, public_key_file, priva
         cert_file, username = _get_and_write_certificate(cmd, public_key_file, None)
         if is_arc:
             try:
-                certificate_validity_in_seconds = _get_certificate_validity_in_seconds(cert_file)
+                cert_lifetime = ssh_utils._get_certificate_lifetime(cert_file).total_seconds()
+                print(cert_lifetime)
             except Exception as e:
-                logger.warning("Couldn't determine certificate validity. Error: %s", str(e))
-            print(certificate_validity_in_seconds)
+                logger.warning("Couldn't determine certificate expiration. Error: %s", str(e))
 
     proxy_path = None
     relay_info = None
     if is_arc:
-        proxy_path = _arc_get_client_side_proxy(arc_proxy_folder)
-        relay_info = _arc_list_access_details(cmd, resource_group_name, vm_name, certificate_validity_in_seconds)
+        proxy_path = arc_utils._arc_get_client_side_proxy(ssh_proxy_folder)
+        relay_info = arc_utils._arc_list_access_details(cmd, resource_group_name, vm_name, cert_lifetime)
     else:
         ssh_ip = ssh_ip or ip_utils.get_ssh_ip(cmd, resource_group_name, vm_name, use_private_ip)
         if not ssh_ip:
@@ -159,18 +159,6 @@ def _do_ssh_op(cmd, vm_name, resource_group_name, ssh_ip, public_key_file, priva
 
     op_call(relay_info, proxy_path, vm_name, ssh_ip, username, cert_file, private_key_file, port, is_arc, delete_keys,
             delete_cert, public_key_file)
-
-
-def _get_certificate_validity_in_seconds(cert_file):
-    import datetime
-    validity_str = ssh_utils.get_ssh_cert_validity(cert_file)
-    validity = None
-    if validity_str and "Valid: from " in validity_str and " to " in validity_str:
-        times = validity_str.replace("Valid: from ", "").split(" to ")
-        t0 = datetime.datetime.fromisoformat(times[0])
-        t1 = datetime.datetime.fromisoformat(times[1])
-        validity = t1 - t0
-    return int(validity.total_seconds())
 
 
 def _get_and_write_certificate(cmd, public_key_file, cert_file):
@@ -320,132 +308,6 @@ def _get_modulus_exponent(public_key_file):
     exponent = parser.exponent
 
     return modulus, exponent
-
-
-# Downloads client side proxy to connect to Arc Connectivity Platform
-def _arc_get_client_side_proxy(arc_proxy_folder):
-
-    request_uri, install_location, older_version_location = _get_proxy_filename_and_url(arc_proxy_folder)
-    install_dir = os.path.dirname(install_location)
-
-    # Only download new proxy if it doesn't exist already
-    if not os.path.isfile(install_location):
-        t0 = time.time()
-        # download the executable
-        try:
-            with urllib.request.urlopen(request_uri) as response:
-                response_content = response.read()
-                response.close()
-        except Exception as e:
-            telemetry.set_exception(exception=e, fault_type=consts.PROXY_DOWNLOAD_FAILED_FAULT_TYPE,
-                                    summary=f'Failed to download proxy from {request_uri}')
-            raise azclierror.ClientRequestError(f"Failed to download client proxy executable from {request_uri}. "
-                                                "Error: " + str(e)) from e
-        time_elapsed = time.time() - t0
-
-        proxy_data = {
-            'Context.Default.AzureCLI.SSHProxyDownloadTime': time_elapsed,
-            'Context.Default.AzureCLI.SSHProxyVersion': consts.CLIENT_PROXY_VERSION
-        }
-        telemetry.add_extension_event('ssh', proxy_data)
-
-        # if directory doesn't exist, create it
-        if not os.path.isdir(install_dir):
-            file_utils.create_directory(install_dir, f"Failed to create client proxy directory '{install_dir}'. ")
-        # if directory exists, delete any older versions of the proxy
-        else:
-            older_version_files = glob(older_version_location)
-            for f in older_version_files:
-                file_utils.delete_file(f, f"failed to delete older version file {f}", warning=True)
-
-        # write executable in the install location
-        file_utils.write_to_file(install_location, 'wb', response_content, "Failed to create client proxy file. ")
-        os.chmod(install_location, os.stat(install_location).st_mode | stat.S_IXUSR)
-
-    return install_location
-
-
-def _get_proxy_filename_and_url(arc_proxy_folder):
-    import platform
-    operating_system = platform.system()
-    machine = platform.machine()
-
-    logger.debug("Platform OS: %s", operating_system)
-    logger.debug("Platform architecture: %s", machine)
-
-    if machine.endswith('64'):
-        architecture = 'amd64'
-    elif machine.endswith('86'):
-        architecture = '386'
-    elif machine == '':
-        raise azclierror.BadRequestError("Couldn't identify the platform architecture.")
-    else:
-        telemetry.set_exception(exception='Unsuported architecture for installing proxy',
-                                fault_type=consts.PROXY_UNSUPPORTED_ARCH_FAULT_TYPE,
-                                summary=f'{machine} is not supported for installing client proxy')
-        raise azclierror.BadRequestError(f"Unsuported architecture: {machine} is not currently supported")
-
-    # define the request url and install location based on the os and architecture
-    proxy_name = f"sshProxy_{operating_system.lower()}_{architecture}"
-    request_uri = (f"{consts.CLIENT_PROXY_STORAGE_URL}/{consts.CLIENT_PROXY_RELEASE}"
-                   f"/{proxy_name}_{consts.CLIENT_PROXY_VERSION}")
-    install_location = proxy_name + "_" + consts.CLIENT_PROXY_VERSION.replace('.', '_')
-    older_location = proxy_name + "*"
-
-    if operating_system == 'Windows':
-        request_uri = request_uri + ".exe"
-        install_location = install_location + ".exe"
-        older_location = older_location + ".exe"
-    elif operating_system not in ('Linux', 'Darwin'):
-        telemetry.set_exception(exception='Unsuported OS for installing ssh client proxy',
-                                fault_type=consts.PROXY_UNSUPPORTED_OS_FAULT_TYPE,
-                                summary=f'{operating_system} is not supported for installing client proxy')
-        raise azclierror.BadRequestError(f"Unsuported OS: {operating_system} platform is not currently supported")
-
-    if not arc_proxy_folder:
-        install_location = os.path.expanduser(os.path.join('~', os.path.join(".clientsshproxy", install_location)))
-        older_location = os.path.expanduser(os.path.join('~', os.path.join(".clientsshproxy", older_location)))
-    else:
-        install_location = os.path.join(arc_proxy_folder, install_location)
-        older_location = os.path.join(arc_proxy_folder, older_location)
-
-    return request_uri, install_location, older_location
-
-
-# Get the Access Details to connect to Arc Connectivity platform from the HybridConnectivity RP
-def _arc_list_access_details(cmd, resource_group, vm_name, certificate_validity):
-    from azext_ssh._client_factory import cf_endpoint
-    client = cf_endpoint(cmd.cli_ctx)
-    try:
-        t0 = time.time()
-        if certificate_validity:
-            result = client.list_credentials(resource_group_name=resource_group, machine_name=vm_name,
-                                             endpoint_name="default", expiresin=certificate_validity)
-        else:
-            result = client.list_credentials(resource_group_name=resource_group, machine_name=vm_name,
-                                             endpoint_name="default")
-        time_elapsed = time.time() - t0
-        telemetry.add_extension_event('ssh', {'Context.Default.AzureCLI.SSHListCredentialsTime': time_elapsed})
-    except Exception as e:
-        telemetry.set_exception(exception='Call to listCredentials failed',
-                                fault_type=consts.LIST_CREDENTIALS_FAILED_FAULT_TYPE,
-                                summary=f'listCredentials failed with error: {str(e)}.')
-        raise azclierror.ClientRequestError(f"Request for Azure Relay Information Failed: {str(e)}")
-
-    result_string = json.dumps(
-        {
-            "relay": {
-                "namespaceName": result.namespace_name,
-                "namespaceNameSuffix": result.namespace_name_suffix,
-                "hybridConnectionName": result.hybrid_connection_name,
-                "accessKey": result.access_key,
-                "expiresOn": result.expires_on
-            }
-        })
-    result_bytes = result_string.encode("ascii")
-    enc = base64.b64encode(result_bytes)
-    base64_result_string = enc.decode("ascii")
-    return base64_result_string
 
 
 def _decide_op_call(cmd, resource_group_name, vm_name, ssh_ip, resource_type, config_path, overwrite,
